Move anaconda env initialization to top of bashrc
bashrc is read at each session start but it has a statement to do
nothing if the session is not interactivetly. This spoiled its usage
when sending jobs to my computing cluster which as a matter of fact is a
non-interactive session. That caused scripts to fail as they could not
find any of the environment variables set for them.

Now the anaconda environment variables are loaded always, if the machine
is within the cluster they are started. Then if the session is
interactive other paths and aliases are loaded.