Relative line numbers per mode
This is because the global option launches also in pdf-view and makes it
completely slow. Same as when loaded with org-mode time ago. So now is
only for text and prog mode. Leaving other modes like pdf, helm,
packages where no line numbers are needed out, and eventually new modes.