Stop using latex toggle automatically
Although it looks like a very nice feature to render latex immediately
after writing it. This features was causing a lot of unwanted
behavior. When entering a backslash point jumps back, In inline math
environment is cumbersome to jump to end of it, and the image pop in
in between the point movement making it very complicated.

Using a keychord to launch the render instead of the crazy emacs
keycombo is faster and for now looks like a better option.