Make agenda keybindings work
For a strange reason the use-package :map was not working with the
startup run of Emacs. But a re-execution of that block did allow for the
moment key to work. As I would never be opening the config just to get
the key binding I put a hook to modify such keys.

Now it works automatically and I don't need to think about another
keyset.