Don't load anaconda and Emacs on system start
I do tend to restart Emacs more often than my average time of an open
session. Thus it makes little sense to have this autostart with anaconda
and emacs I can do it manually.

Second a better way to use the anaconda environment is to use pyvenv,
and not this launch script