Use default org-export-babel-evaluate t
The default is like that for reproducibility reasons. I avoided it,
because some of my python code was to time consuming to execute on
exports. Now I avoid evaluation on export as a header flag which is in
place by default.

The reason for this change, probably Org-mode 9. When not evaluating
on export, no header was obeyed and code and output where exported. I
don't want my code to be in the output.